refactor: features/themeManagement 제거
