--- frontend/techpick/src/features/index.ts
@@ -1,1 +1,0 @@
-export { ToggleThemeButton } from '@/features/themeManagement/ToggleThemeButton';
